Tidy up JEI integration
