+++ dependencies.properties
@@ -1,2 +1,2 @@
 forge=47.1.3
-jei=4712868
+jei=15.17.0.76
